refactor: move ai files under ai folder and update paths
